refactor: moved emotion component to separate file
